requirements.txt
xlwings
ajax==3.0.0
alabaster==0.7.12
appdirs==1.4.3
argh==0.26.2
asgiref==3.2.10
asn1crypto==1.3.0
astroid==2.3.3
astropy==3.2.3
atomicwrites==1.4.0
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.7
beautifulsoup4==4.9.1
bitarray==1.2.2
bkcharts==0.2
bleach==3.1.5
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
cssselect==1.1.0
cssutils==1.0.2
cycler==0.10.0
Cython==0.29.20
cytoolz==0.10.1
decorator==4.4.2
defusedxml==0.5.0
diff-match-patch==20181111
distlib==0.3.0
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.37.1
django-appconf==1.0.3
django-ckeditor==6.0.0
django-cors-headers==2.4.0
django-crispy-forms==1.9.0
django-csvimport==2.15
django-extensions==2.2.8
django-filter==2.3.0
django-geojson==2.12.0
django-imagekit==4.0.2
django-import-export==2.0.2
django-jquery==3.1.0
django-js-asset==1.2.2
django-json==0.27
django-jsonfield==1.4.0
django-leaflet==0.26.0
django-modeltranslation==0.15.2
django-mysql==3.6.0
django-pandas==0.6.1
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-send-email==0.1.0
django-simple-mail==2.3.3
django-tables2==2.3.1
django-widget-tweaks==1.4.1
djangorestframework==3.8.2
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
fsspec==0.7.4
future==0.18.2
geographiclib==1.50
geojson==2.5.0
geopy==1.21.0
glob2==0.7

chartjs==1.2
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.3
comtypes==1.1.7
contextlib2==0.6.0.post1

greenlet==0.4.16
gunicorn==20.0.4
h5py==2.10.0
HeapDict==1.0.1
html2text==2020.1.16
html5lib==1.0.1
idna==2.9
imageio==2.8.0
imagesize==1.2.0
importlib-metadata==1.6.1
intervaltree==3.0.2
ipykernel==5.3.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
joblib==0.15.1
json-table-schema==0.2.1
json5==0.9.5
jsonfield==2.0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.1.4
jupyterlab-server==1.1.5
kiwisolver==1.2.0
leaflet==0.0.3
libarchive-c==2.9
locket==0.2.0
lxml==4.5.1
Markdown==3.2.2
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.3.0rc1
mccabe==0.6.1
messytables==0.15.2
mistune==0.8.4

mock==4.0.2
more-itertools==8.4.0
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0

nbconvert==5.6.1
nbformat==5.0.7
nbopen==0.6
networkx==2.4
nose==1.3.7
notebook==6.0.3
numba==0.50.1
numexpr==2.7.1
numpy==1.18.5
numpydoc==1.0.0
oauthlib==2.1.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.3
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==7.1.2
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
premailer==3.7.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psutil==5.7.0
psycopg2==2.8.5
ptyprocess==0.6.0
py==1.8.2
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20

pycurl==7.43.0.5
pyflakes==2.2.0
pygeocoder==1.2.5
Pygments==2.6.1
pylint==2.4.4



pyOpenSSL==19.1.0
pyparsing==2.4.7
pyproj==1.9.6
pyreadline==2.1
pyrsistent==0.16.0
pyshp==2.1.0
PySocks==1.7.1
pytest==5.4.3
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.7.0
pytest-openfiles==0.5.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-decouple==3.1
python-dotenv==0.14.0
python-jsonrpc-server==0.3.4
python-magic==0.4.18
python3-openid==3.1.0
PyWavelets==1.1.1
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
QtPy==1.9.0

requests-oauthlib==1.0.0
reverse-geocoder==1.5.1
rope==0.17.0


runipy==0.1.5
scikit-image==0.16.2
scikit-learn==0.20.0

seaborn==0.10.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1

sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4

spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.17
sqlparse==0.3.0
statsmodels==0.11.1
sympy==1.6
tables==3.6.1
tablib==1.1.0
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.2
UgliPyJS==0.2.5
ujson
unicodecsv==0.14.1
Unipath==1.1
urllib3==1.25.9
virtualenv==20.0.9
virtualenvwrapper-win==1.2.6
watchdog==0.10.2
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
xlrd==1.2.0
XlsxWriter==1.2.9
xlwt==1.3.0
xmltodict==0.12.0
zict==2.0.0
zipp==3.1.0
zope.event==4.4
zope.interface==4.7.1
